Remove song alias commands
